+++ src/main/scala/com/scalakml/example/WriteExample4.scala
@@ -12,7 +12,7 @@ import com.scalakml.kml._
 
 
 object WriteExample4 {
-  def main(args: Array[String])  {
+  def main(args: Array[String]): Unit = {
     // create a scala Kml object with a Placemark that contains a Point
     val kml = new Kml(new Placemark("Sydney", new Point(RelativeToGround, 151.21037, -33.8526, 12345.0)))
     // write the kml object to System.out as xml